@@ -2,7 +2,7 @@ package com.digitalaicloud.hibernate.entity;
 
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @NoArgsConstructor
