@@ -36,7 +36,7 @@ public class HibernateApplication implements CommandLineRunner {
 //		courseRepository.playWithEntityManager();
 
 //		courseRepository.playWithEntityManagerUpdateFalse();
-//		courseRepository.playWithEntityManagerInsertFalse();
+		courseRepository.playWithEntityManagerInsertFalse();
 	}
 
 
